refactor: remove obsolete websocket handler
